+++ src/app/tasks-page/tasks-page.tsx
@@ -2,25 +2,23 @@ import React, { FC, useState } from 'react';
 import { Button, Col } from 'react-bootstrap';
 import CharacterList from '../character-list/character-list';
 import CreateCharacterModal from '../create-character-modal/create-character-modal';
-import styles from './tasks.module.css';
+import styles from './tasks-page.module.css';
 
-interface TasksProps { }
+interface TasksPageProps { }
 
-const Tasks: FC<TasksProps> = () => {
+const TasksPage: FC<TasksPageProps> = () => {
 
   const [show, setShow] = useState(false);
   const handleCloseCharacterCreationModal = () => setShow(false);
   const showCharacterCreationModal = () => setShow(true);
 
   return (
-    <div className={styles.Tasks} data-testid="Tasks">
+    <div className={styles.tasksPage} data-testid="tasksPage">
       <Button onClick={showCharacterCreationModal}> Add character </Button>
       <CreateCharacterModal show={show} close={handleCloseCharacterCreationModal}></CreateCharacterModal>
-      <Col sm={2}>
-        <CharacterList></CharacterList>
-      </Col>
+      <CharacterList></CharacterList>
     </div >
   )
 };
 
-export default Tasks;
+export default TasksPage;
