--- src/app/tasks/tasks.module.css
@@ -1,1 +1,0 @@
-.Tasks {}
